Add players using a form
We add a simple form to add player names associated with a color on the game. This form is used to try to create a new `Players` value, which can only be created if at least two players exist. If we  fail to create the new value, we recursively render the form again with the last known state and let the user add more players. 

Up to four players can be added to the game.

Some of the changes in this branch relate to "walking" players on the board, a feature that will be completed on a different feature branch.

* Rename some fields in Game
* Reorder Color enum to match board placement
* Remove `Check` phase: `Check` is just a transition between `Walk` and `End`
* Rename Player back to Players
* Add scaffolding for user creation form
* Add players using a form
* Remove all redundant imports