Refactor Players and tokens
Instead of using a custom data type for Players, we're using a map instedad, and
hiding it behind a newtype. We then provide a single way to create a Players map
to prevent any Game with unsound numbers of players.

During game play, we never move the players around anymore, but instead we move
tokens representing players on the board. Tokens are just Color placed in each
tile.

Finally, instead of storing the treasure information in the Player directly, we
create another map in which the list of treasures for a given color can be
retrieved by using the token Color for such player.